Update django-wiki version for pip install
When I ran `pip install -r requirements.txt`, I got a weird error that it
couldn't find a version/package matching `wiki==0.0.24`, but on the repo page
https://github.com/django-wiki/django-wiki and in the same error printout, it
indicated that 0.0.24.1 , 0.0.24.2, and 0.0.24.3 were acceptable versions. I
changed the entry to the most recent one and now the pip install works (tested
it, obvi).

Relatedly, as their page indicates that "django-wiki 0.1 is right on the
doorstep", we may want to switch to it soon when it comes out.

Changes to be committed:
	modified:   requirements.txt